+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add functions to access inner values
+    - `Service.domain`
+    - `Service.context`
+    - `Processor.purpose`
+    - `Instance.version`
+    - `Spot.zone`
+    - `Spot.bucket`
 
 ## 3.1.0 - 2019-08-21
 - Add more create functions